// src/interfaces/predial.interface.ts
export interface Predial {

    uuid?: string;
    fecha_oficio?: string;
    numero_oficio?: string;
    recibo_pago_certificado?: string;
    folio_pago_predial?: string;
    clave_catastral?: string;
    expediente?: string;
    ultimo_bimestre_ejercicio_pagado?: string;
    ubicacion?: string;
    nombre_contribuyente?: string;
    rfc?: string;

}